Extract retry/rate-limit orchestration from use-chat
use-chat.ts had grown to 392 lines (the per-file limit), with the
80-line executeWithRetry callback handling rate-limit detection,
exponential backoff with abort-aware delays, and per-step autosave
bookkeeping inline alongside conversation/message state. That pushed
it against the file limit and made the retry behavior hard to test
in isolation.

Move the executeWithRetry callback into a new useExecuteWithRetry
hook in helpers/. The hook owns its own retry AbortController and
exposes an abortRetry() canceler so clearConversation/stopResponse
can still cancel a pending backoff. use-chat now calls the hook with
its message setters and shrinks to 312 lines.

Behavior is unchanged — all 21 existing retry tests pass against the
extracted hook and 169 chat-hook tests pass overall.
